chore: make blob params u64

+++ crates/context/interface/src/cfg.rs
@@ -15,7 +15,7 @@ pub trait Cfg {
     /// Returns the blob target and max count for the given spec id.
     ///
     /// EIP-7840: Add blob schedule to execution client configuration files
-    fn blob_max_count(&self, spec_id: SpecId) -> u8;
+    fn blob_max_count(&self, spec_id: SpecId) -> u64;
 
     fn max_code_size(&self) -> usize;
